feat: initial iOS support (#169)
- closes https://github.com/josxha/flutter-maplibre/issues/26

Check the tasklist of the issue to follow the current progress.

<details>
<summary>Outdated information</summary>


## Licensing

> [!WARNING]
> The iOS implementation is during the funding phase provided under
> the [Non-Profit Open Software License version 3.0
(NPOSL-3.0)](https://opensource.org/license/osl-3-0-php).
> After funding is completed, it will get released under the permissive
3-Clause BSD License.
> For more information, see:
https://github.com/josxha/flutter-maplibre/pull/169#issuecomment-2601023222

There was a question with my answer about what what "funding phase"
means, that is now hidden in all the commits of this pr:

> iOS support is currently the most requested feature and the biggest
blockers of using this package.
As I want to offer support for web, android and ios, I'll happily add
support for iOS. Though, this required me to purchase new up-to-date
apple hardware. To guarantee the ability for long term maintainence, I
went for an mac mini m4 with magic keyboard. I'd like to add an ipad to
the list to test on an physical device when the implementation is about
to be finished (in total ~ 1.330€ or 1.367$).
> 
> I donate my time and effort to open source, but am not able to carry
these costs for an open source project. Even though, I pre-funded the
hardware and started implementation so that sponsors have a decent
guarantee of delivery. The only use those devices had since I bought
them is the development of the maplibre flutter package - so there is
really no dual use here. 😄
> 
> The idea is to do a crowd-funding campain so that multiple parties can
fund this feature together. There has been some interest in sponsoring,
for example in
https://github.com/josxha/flutter-maplibre/issues/111#issue-2611425339.
I want to wait until February when I have more time to annouce the
campain and (hopefully) finish the implementation.
> 
> In the meantime, the iOS implementation gets published in this pull
request under the OSI approved "The Open Software License 3.0". This
allows commercial interested parties to follow along its progress and
non profit projects to already get their hands on early.
> As soon as the hardware costs are funded, iOS support gets released
under the permissive BSD-3-Clause license that allows commercial use as
it is not in my interest to maintain two variants of the package.
> 
> I really do hope that this is in the interest of everyone and I am
open for any feedback or suggestions.


</details>

## How to use

[Turn on the Swift Package
Manager](https://docs.flutter.dev/packages-and-plugins/swift-package-manager/for-app-developers#how-to-turn-on-swift-package-manager)
to be able to run the app:
```console
flutter config --enable-swift-package-manager
```

### Task list (WIP)

- [x] #163
- [x] #166
- [x] Add ios CI builds,
https://github.com/josxha/flutter-maplibre/pull/170
- [x] Check Swift formatting and lints in CI,
https://github.com/josxha/flutter-maplibre/pull/171
- [x] Access the native map view instance from dart,
https://github.com/josxha/flutter-maplibre/pull/172
- [x] Programmatic map controls,
https://github.com/josxha/flutter-maplibre/pull/172
- [x] Apply paint and layout properties,
https://github.com/josxha/flutter-maplibre/pull/175
- [x] Apply initial MapOptions
- [x] ensure compatebility with the stable Flutter channel
- [x] Make it work with Flutter state management
- [x] ~~PermissionManager~~, does not exist on ios
- [x] OfflineManager & cache
- [ ] Encode offline region data into the metadata for partiy with
android
- [ ] Add ios integration test setup
- [x] ~~Use pub workspace
https://medium.com/dartlang/announcing-dart-3-6-778dd7a80983~~, Refactor
package after this pr has been merged.
- [x] ~~Adjust pub.dev release process~~, need to be updated once an
initial version has been published.
- [x] Get Github Actions to run sucessfully
- [ ] apply initial camera pitch
- [x] fix: interactive marker example can't pan map
- [x] do not show PermissionManager example page on iOS
- [ ] fix: crash on Marker Layer example page
- [ ] fix: crash on Polylines Layer example page