Revert "use arena for ios implementation"
This reverts commit 8432435d40d7df82974fac46c04c8e161522ff64.

from the docs:

References are automatically retained when the Dart wrapper object is created, and released when the Dart object is garbage collected. But to fix a reference counting issue, it may occasionally be necessary to manually control the ref count. The ref count of the Objective C object can be controlled from Dart by the retain and release options in the Dart wrapper object constructors, and by the retainAndReturnPointer() and release() methods.